chore: move more common editor code
